backend/pyproject.toml
[project]
name = "woc-backend"
version = "0.1.0"
description = "Next gen backend of worldofcode.org"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "beanie>=1.29.0",
    "clickhouse-driver>=0.2.9",
    "dynaconf>=3.2.7",
    "fastapi>=0.115.8",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "python-woc>=0.2.5",
    "shortuuid>=1.0.13",
    "uvicorn[standard]>=0.33.0",
    "aiosmtplib>=3.0.2,<4",
    "jinja2>=3.1.6",
    "python-multipart>=0.0.20",
]

[dependency-groups]
dev = [
    "mypy>=1.14.1",
    "pytest>=8.3.4",
    "ruff>=0.9.4",
]
